// src/models/logtools.ts
export type Logtools = {
 title: string;
 sub_content: string;
 achieve_content: string;
 content: string;
 play_item: number;
 clear_item: number;
 created_at: string;
 created_by: string;
 updated_at: string;
}

export type PostLogtools = {
 title: string,
 sub_content: string,
 achieve_content: string,
 content: string,
 play_item: number,
 clear_item: number,
}